Shortened and refactored batch methods
