+++ Api/Api.csproj
@@ -36,11 +36,11 @@
     <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore" Version="3.1.2" />
     <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="3.1.2" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.1" />
-    <PackageReference Include="OpenTelemetry" Version="0.2.0-alpha.277" />
-    <PackageReference Include="OpenTelemetry.Adapter.AspNetCore" Version="0.2.0-alpha.277" />
-    <PackageReference Include="OpenTelemetry.Adapter.Dependencies" Version="0.2.0-alpha.277" />
-    <PackageReference Include="OpenTelemetry.Exporter.Jaeger" Version="0.2.0-alpha.277" />
-    <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="0.2.0-alpha.277" />
+    <PackageReference Include="OpenTelemetry" Version="0.2.0-alpha.275" />
+    <PackageReference Include="OpenTelemetry.Adapter.AspNetCore" Version="0.2.0-alpha.275" />
+    <PackageReference Include="OpenTelemetry.Adapter.Dependencies" Version="0.2.0-alpha.275" />
+    <PackageReference Include="OpenTelemetry.Exporter.Jaeger" Version="0.2.0-alpha.275" />
+    <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="0.2.0-alpha.275" />
     <PackageReference Include="Sendgrid" Version="9.12.6" />
     <PackageReference Include="Sentry.AspNetCore" Version="2.1.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0" />
