+++ Api/HappyTravel.Edo.Api.xml
@@ -344,21 +344,28 @@
             <param name="bookingIds">List of booking ids for cancellation</param>
             <returns>Result message</returns>
         </member>
-        <member name="M:HappyTravel.Edo.Api.Controllers.InternalPaymentsController.GetBookingsForCapture(System.DateTime)">
+        <member name="M:HappyTravel.Edo.Api.Controllers.InternalBookingsController.GetBookingsForCapture(System.DateTime)">
             <summary>
                 Gets bookings for payment completion by deadline date
             </summary>
             <param name="deadlineDate">Deadline date</param>
             <returns>List of booking ids for capture</returns>
         </member>
-        <member name="M:HappyTravel.Edo.Api.Controllers.InternalPaymentsController.Capture(System.Collections.Generic.List{System.Int32})">
+        <member name="M:HappyTravel.Edo.Api.Controllers.InternalBookingsController.Capture(System.Collections.Generic.List{System.Int32})">
             <summary>
                 Captures payments for bookings
             </summary>
             <param name="bookingIds">List of booking ids for capture</param>
             <returns>Result message</returns>
         </member>
-        <member name="M:HappyTravel.Edo.Api.Controllers.InternalPaymentsController.NotifyPaymentsNeeded(System.Collections.Generic.List{System.Int32})">
+        <member name="M:HappyTravel.Edo.Api.Controllers.InternalBookingsController.GetBookingsToNotify(System.DateTime)">
+            <summary>
+                Sends need payment notifications for bookings
+            </summary>
+            <param name="deadlineDate">Deadline date</param>
+            <returns>Result message</returns>
+        </member>
+        <member name="M:HappyTravel.Edo.Api.Controllers.InternalBookingsController.NotifyPaymentsNeeded(System.Collections.Generic.List{System.Int32})">
             <summary>
                 Sends need payment notifications for bookings
             </summary>
