+++ Api/Services/Accommodations/Bookings/BookingJobsService.cs
@@ -1,0 +1,229 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Text;
+using System.Threading.Tasks;
+using CSharpFunctionalExtensions;
+using HappyTravel.Edo.Api.Infrastructure;
+using HappyTravel.Edo.Api.Infrastructure.DataProviders;
+using HappyTravel.Edo.Api.Models.Bookings;
+using HappyTravel.Edo.Api.Models.Payments;
+using HappyTravel.Edo.Api.Models.Users;
+using HappyTravel.Edo.Api.Services.Payments;
+using HappyTravel.Edo.Common.Enums;
+using HappyTravel.Edo.Data;
+using HappyTravel.Edo.Data.Booking;
+using HappyTravel.Edo.Data.Management;
+using HappyTravel.EdoContracts.Accommodations.Enums;
+using HappyTravel.EdoContracts.General.Enums;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+
+namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
+{
+    public class BookingJobsService : IBookingJobsService
+    {
+        public BookingJobsService(IBookingPaymentService bookingPaymentService,
+            IPaymentNotificationService notificationService,
+            IBookingService bookingService,
+            IDateTimeProvider dateTimeProvider,
+            EdoContext context)
+        {
+            _bookingPaymentService = bookingPaymentService;
+            _notificationService = notificationService;
+            _bookingService = bookingService;
+            _dateTimeProvider = dateTimeProvider;
+            _context = context;
+        }
+
+
+        public async Task<Result<List<int>>> GetForCapture(DateTime deadlineDate)
+        {
+            if (deadlineDate == default)
+                return Result.Failure<List<int>>("Deadline date should be specified");
+
+            var date = deadlineDate.Date;
+
+            var bookingIds = await _context.Bookings
+                .Where(IsBookingValidForCapturePredicate)
+                .Where(b => b.CheckInDate <= date || b.DeadlineDate.HasValue && b.DeadlineDate.Value.Date < date)
+                .Select(b => b.Id)
+                .ToListAsync();
+
+            return Result.Ok(bookingIds);
+        }
+
+
+        public Task<Result<ProcessResult>> Capture(List<int> bookingIds, ServiceAccount serviceAccount)
+        {
+            return ExecuteBatchJob(bookingIds,
+                IsBookingValidForCapturePredicate,
+                Capture,
+                serviceAccount);
+
+            Task<Result<string>> Capture(Booking booking, UserInfo serviceAcc) => _bookingPaymentService.CaptureMoney(booking, serviceAccount.ToUserInfo());
+        }
+        
+        
+        public Task<Result<List<int>>> GetForNotify(DateTime deadlineDate) => GetForCapture(deadlineDate.AddDays(DaysBeforeNotification));
+
+
+        public Task<Result<ProcessResult>> NotifyDeadlineApproaching(List<int> bookingIds, ServiceAccount serviceAccount)
+        {
+            return ExecuteBatchJob(bookingIds,
+                IsBookingValidForCapturePredicate,
+                Notify,
+                serviceAccount);
+
+
+            Task<Result<string>> Notify(Booking booking, UserInfo _)
+            {
+                return Notify()
+                    .Finally(CreateResult);
+
+
+                async Task<Result> Notify()
+                {
+                    var agent = await _context.Agents.SingleOrDefaultAsync(a => a.Id == booking.AgentId);
+                    if (agent == default)
+                        return Result.Failure($"Could not find agent with id {booking.AgentId}");
+
+                    return await _notificationService.SendNeedPaymentNotificationToCustomer(new PaymentBill(agent.Email,
+                        booking.TotalPrice,
+                        booking.Currency,
+                        DateTime.MinValue,
+                        booking.PaymentMethod,
+                        booking.ReferenceCode,
+                        $"{agent.LastName} {agent.FirstName}"));
+                }
+
+
+                Result<string> CreateResult(Result result)
+                    => result.IsSuccess
+                        ? Result.Ok($"Payment for the booking '{booking.ReferenceCode}' completed.")
+                        : Result.Failure<string>($"Unable to complete payment for the booking '{booking.ReferenceCode}'. Reason: {result.Error}");
+            }
+        }
+
+
+        public async Task<Result<List<int>>> GetForCancellation(DateTime deadlineDate)
+        {
+            if (deadlineDate == default)
+                return Result.Failure<List<int>>("Deadline date should be specified");
+
+            // It’s prohibited to cancel booking after check-in date
+            var currentDateUtc = _dateTimeProvider.UtcNow();
+            var dayBeforeDeadline = deadlineDate.Date.AddDays(1);
+            var bookingIds = await _context.Bookings
+                .Where(booking =>
+                    booking.BookingDate > currentDateUtc &&
+                    booking.DeadlineDate != null &&
+                    booking.DeadlineDate.Value <= dayBeforeDeadline
+                )
+                .Where(IsBookingValidForCancelPredicate)
+                .Select(booking => booking.Id)
+                .ToListAsync();
+
+            return Result.Ok(bookingIds);
+        }
+
+
+        public Task<Result<ProcessResult>> Cancel(List<int> bookingIds, ServiceAccount serviceAccount)
+        {
+            return ExecuteBatchJob(bookingIds,
+                IsBookingValidForCancelPredicate,
+                ProcessBooking,
+                serviceAccount);
+
+
+            Task<Result<string>> ProcessBooking(Booking booking, UserInfo _)
+            {
+                return _bookingService.Cancel(booking.Id, serviceAccount)
+                    .Finally(CreateResult);
+
+
+                Result<string> CreateResult(Result<VoidObject, ProblemDetails> result)
+                    => result.IsSuccess
+                        ? Result.Ok($"Booking '{booking.ReferenceCode}' was cancelled.")
+                        : Result.Failure<string>($"Unable to cancel booking '{booking.ReferenceCode}'. Reason: {result.Error.Detail}");
+            }
+        }
+
+
+        private async Task<Result<ProcessResult>> ExecuteBatchJob(List<int> bookingIds,
+            Expression<Func<Booking, bool>> predicate,
+            Func<Booking, UserInfo, Task<Result<string>>> action,
+            ServiceAccount serviceAccount)
+        {
+            var bookings = await GetBookings();
+
+            return await ValidateCount()
+                .Map(ProcessBookings);
+
+
+            Task<List<Booking>> GetBookings()
+                => _context.Bookings
+                    .Where(booking => bookingIds.Contains(booking.Id))
+                    .Where(predicate)
+                    .ToListAsync();
+
+
+            Result ValidateCount()
+                => bookings.Count != bookingIds.Count
+                    ? Result.Failure("Invalid booking ids. Could not find some of requested bookings.")
+                    : Result.Ok();
+
+
+            Task<ProcessResult> ProcessBookings() => Combine(bookings.Select(booking => action(booking, serviceAccount.ToUserInfo())));
+
+
+            async Task<ProcessResult> Combine(IEnumerable<Task<Result<string>>> results)
+            {
+                var builder = new StringBuilder();
+
+                foreach (var result in results)
+                {
+                    var (_, isFailure, value, error) = await result;
+                    builder.AppendLine(isFailure ? error : value);
+                }
+
+                return new ProcessResult(builder.ToString());
+            }
+        }
+
+
+        private const int DaysBeforeNotification = 3;
+
+        private static readonly Expression<Func<Booking, bool>> IsBookingValidForCancelPredicate = booking
+            => BookingStatusesForPayment.Contains(booking.Status) &&
+            booking.PaymentStatus == BookingPaymentStatuses.Authorized &&
+            (booking.PaymentMethod == PaymentMethods.BankTransfer || booking.PaymentMethod == PaymentMethods.CreditCard);
+
+        private static readonly Expression<Func<Booking, bool>> IsBookingValidForCapturePredicate = booking
+            => BookingStatusesForCancellation.Contains(booking.Status) &&
+            PaymentStatusesForCancellation.Contains(booking.PaymentStatus);
+
+        private static readonly HashSet<BookingStatusCodes> BookingStatusesForPayment = new HashSet<BookingStatusCodes>
+        {
+            BookingStatusCodes.Pending, BookingStatusCodes.Confirmed
+        };
+
+        private static readonly HashSet<BookingStatusCodes> BookingStatusesForCancellation = new HashSet<BookingStatusCodes>
+        {
+            BookingStatusCodes.Pending, BookingStatusCodes.Confirmed
+        };
+
+
+        private static readonly HashSet<BookingPaymentStatuses> PaymentStatusesForCancellation = new HashSet<BookingPaymentStatuses>
+        {
+            BookingPaymentStatuses.NotPaid, BookingPaymentStatuses.Authorized
+        };
+
+        private readonly IBookingPaymentService _bookingPaymentService;
+        private readonly IBookingService _bookingService;
+        private readonly EdoContext _context;
+        private readonly IDateTimeProvider _dateTimeProvider;
+        private readonly IPaymentNotificationService _notificationService;
+    }
+}
