+++ Api/Services/Accommodations/Bookings/IBookingJobsService.cs
@@ -7,8 +7,16 @@ using HappyTravel.Edo.Data.Management;
 
 namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
 {
-    public interface IBookingsProcessingService
+    public interface IBookingJobsService
     {
+        Task<Result<List<int>>> GetForCapture(DateTime deadlineDate);
+
+        Task<Result<ProcessResult>> Capture(List<int> bookingIds, ServiceAccount serviceAccount);
+        
+        Task<Result<List<int>>> GetForNotify(DateTime deadlineDate);
+
+        Task<Result<ProcessResult>> NotifyDeadlineApproaching(List<int> bookingIds, ServiceAccount serviceAccount);
+
         Task<Result<List<int>>> GetForCancellation(DateTime deadlineDate);
 
         Task<Result<ProcessResult>> Cancel(List<int> bookingIds, ServiceAccount serviceAccount);
