+++ Api/Services/Accommodations/Bookings/IBookingPaymentService.cs
@@ -1,8 +1,5 @@
-using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.Models.Bookings;
 using HappyTravel.Edo.Api.Models.Users;
 using HappyTravel.Edo.Api.Services.Payments;
 using HappyTravel.Edo.Data.Booking;
@@ -12,14 +9,10 @@ namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
 {
     public interface IBookingPaymentService : IPaymentsService
     {
-        Task<Result<List<int>>> GetBookingsForCapture(DateTime deadlineDate);
-
-        Task<Result<ProcessResult>> CaptureMoneyForBookings(List<int> bookingIds, ServiceAccount serviceAccount);
+        Task<Result<string>> CaptureMoney(Booking booking, UserInfo user);
 
         Task<Result> VoidMoney(Booking booking, UserInfo user);
 
         Task<Result> CompleteOffline(int bookingId, Administrator administratorContext);
-
-        Task<Result<ProcessResult>> NotifyPaymentsNeeded(List<int> bookingIds);
     }
 }
